added player status getter

+++ Source/MyProject/HealthComponent.cpp
@@ -22,7 +22,7 @@ void UHealthComponent::TakeDamage(FDamageData DamageData)
 	}
 }
 
-float UHealthComponent::GetHealth() const
+int UHealthComponent::GetHealth() const
 {
 	return CurrentHealth;
 }
@@ -32,7 +32,7 @@ float UHealthComponent::GetHealthState() const
 	return CurrentHealth / MaxHealth;
 }
 
-void UHealthComponent::AddHealth(float AddiditionalHealthValue)
+void UHealthComponent::AddHealth(int AddiditionalHealthValue)
 {
 	CurrentHealth += AddiditionalHealthValue;
 	if (CurrentHealth > MaxHealth)
