+++ Source/MyProject/MachineryActor.cpp
@@ -88,7 +88,7 @@ void AMachineryActor::Destroyed()
 		Cannon->Destroy();
 }
 
-void AMachineryActor::DamageTaked(float DamageValue)
+void AMachineryActor::DamageTaked(int DamageValue)
 {
 	UE_LOG(LogTemp, Warning, TEXT("%s taked damage:%f Health:%f"), *GetName(), DamageValue, HealthComponent->GetHealth());
 }
